BACKEND/requirements-dev.txt
# FastAPI core
fastapi==0.104.1
uvicorn==0.23.2
pydantic==1.10.13
starlette==0.27.0
python-multipart==0.0.9     # File uploads
psycopg2-binary==2.9.9
 
# USFM parsing with grammar
usfm-grammar==3.1.2
tree-sitter==0.22.3
tree-sitter-usfm3==3.1.2
 
# HTTP requests for Vachan API
httpx==0.27.0
requests==2.31.0
 
# Database
SQLAlchemy==2.0.30
alembic==1.13.1                    # For migrations
 
# File and report generation
openpyxl==3.1.2               # Excel reports
pandas==2.2.2                  # Optional: for manipulating data in reports
 
# Tokenization, string handling
regex==2024.4.16
unidecode==1.3.8
indic-transliteration==2.3.43      # Optional: for Indian script handling
 
# Environment & logging
python-dotenv   ==1.0.1           # For config variables
loguru==0.7.2                    # Better logging


# pytest
pytest-8.4.1-py3
